requirements.txt
# Core dependencies
anthropic>=0.11.0
pypdf>=3.15.0
langchain>=0.0.267
chromadb>=0.4.18
sentence-transformers>=2.2.2
streamlit>=1.26.0

# LLM dependencies with GPU support
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.22.0
bitsandbytes>=0.40.0
optimum>=1.13.0
peft>=0.5.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0